优化部分单元测试代码，以规避潜在的 bug
com.dwarfeng.statistics.impl.service.DriverInfoMaintainServiceImplTest
com.dwarfeng.statistics.impl.service.DriverSupportMaintainServiceImplTest
com.dwarfeng.statistics.impl.service.FilterInfoMaintainServiceImplTest
com.dwarfeng.statistics.impl.service.FilterSupportMaintainServiceImplTest
com.dwarfeng.statistics.impl.service.HistoryTaskEventMaintainServiceImplTest
com.dwarfeng.statistics.impl.service.HistoryTaskMaintainServiceImplTest
com.dwarfeng.statistics.impl.service.MapperSupportMaintainServiceImplTest
com.dwarfeng.statistics.impl.service.ProviderInfoMaintainServiceImplTest
com.dwarfeng.statistics.impl.service.ProviderSupportMaintainServiceImplTest
com.dwarfeng.statistics.impl.service.StatisticsExecutionProfileMaintainServiceImplTest
com.dwarfeng.statistics.impl.service.StatisticsSettingMaintainServiceImplTest
com.dwarfeng.statistics.impl.service.TaskEventMaintainServiceImplTest
com.dwarfeng.statistics.impl.service.TaskMaintainServiceImplTest
com.dwarfeng.statistics.impl.service.VariableMaintainServiceImplTest

+++ statistics-impl/src/test/java/com/dwarfeng/statistics/impl/service/DriverInfoMaintainServiceImplTest.java
@@ -15,6 +15,7 @@ import org.springframework.test.context.junit4.SpringJUnit4ClassRunner;
 
 import java.util.ArrayList;
 import java.util.List;
+import java.util.Objects;
 
 import static org.junit.Assert.assertEquals;
 import static org.junit.Assert.assertFalse;
@@ -59,9 +60,14 @@ public class DriverInfoMaintainServiceImplTest {
             }
         } finally {
             for (DriverInfo driverInfo : driverInfos) {
+                if (Objects.isNull(driverInfo.getKey())) {
+                    continue;
+                }
                 driverInfoMaintainService.deleteIfExists(driverInfo.getKey());
             }
-            statisticsSettingMaintainService.deleteIfExists(statisticsSetting.getKey());
+            if (Objects.nonNull(statisticsSetting.getKey())) {
+                statisticsSettingMaintainService.deleteIfExists(statisticsSetting.getKey());
+            }
         }
     }
 
@@ -97,9 +103,14 @@ public class DriverInfoMaintainServiceImplTest {
             }
         } finally {
             for (DriverInfo driverInfo : driverInfos) {
+                if (Objects.isNull(driverInfo.getKey())) {
+                    continue;
+                }
                 driverInfoMaintainService.deleteIfExists(driverInfo.getKey());
             }
-            statisticsSettingMaintainService.deleteIfExists(statisticsSetting.getKey());
+            if (Objects.nonNull(statisticsSetting.getKey())) {
+                statisticsSettingMaintainService.deleteIfExists(statisticsSetting.getKey());
+            }
         }
     }
 }
