+++ statistics-impl/src/test/java/com/dwarfeng/statistics/impl/service/HistoryTaskMaintainServiceImplTest.java
@@ -16,6 +16,7 @@ import org.springframework.test.context.junit4.SpringJUnit4ClassRunner;
 import java.util.ArrayList;
 import java.util.Date;
 import java.util.List;
+import java.util.Objects;
 
 import static org.junit.Assert.assertEquals;
 import static org.junit.Assert.assertFalse;
@@ -63,9 +64,14 @@ public class HistoryTaskMaintainServiceImplTest {
             }
         } finally {
             for (HistoryTask historyTask : historyTasks) {
+                if (Objects.isNull(historyTask.getKey())) {
+                    continue;
+                }
                 historyTaskMaintainService.deleteIfExists(historyTask.getKey());
             }
-            statisticsSettingMaintainService.deleteIfExists(statisticsSetting.getKey());
+            if (Objects.nonNull(statisticsSetting.getKey())) {
+                statisticsSettingMaintainService.deleteIfExists(statisticsSetting.getKey());
+            }
         }
     }
 
@@ -101,9 +107,14 @@ public class HistoryTaskMaintainServiceImplTest {
             }
         } finally {
             for (HistoryTask historyTask : historyTasks) {
+                if (Objects.isNull(historyTask.getKey())) {
+                    continue;
+                }
                 historyTaskMaintainService.deleteIfExists(historyTask.getKey());
             }
-            statisticsSettingMaintainService.deleteIfExists(statisticsSetting.getKey());
+            if (Objects.nonNull(statisticsSetting.getKey())) {
+                statisticsSettingMaintainService.deleteIfExists(statisticsSetting.getKey());
+            }
         }
     }
 }
